src/requirements.txt
psycopg2-binary==2.8.6
Django==3.0.14
django-cors-headers==3.11.0
djangorestframework==3.13.1
dj-database-url==0.5.0
numpy==2.1.1
jsonfield==3.1.0
SQLAlchemy==1.4.41
gunicorn==20.1.0
django-environ==0.11.2
requests==2.31.0